placing self service order
