+++ src/Template/Layout/default.ctp
@@ -45,6 +45,7 @@ echo $this->element('layout/header');
     <div id="right">
         <div class="inner-right">
             <?php echo $this->element('cart', [
+                'selfServiceModeEnabled' => false,
                 'showLinkToSelfService' => true,
                 'showLoadLastOrderDetailsDropdown' => true,
                 'showCartDetailButton' => true,
